updated the code to test with new datasets

+++ ingestion/db_scripts/donor.sql
@@ -2,7 +2,7 @@ CREATE TABLE donor (
   id serial PRIMARY KEY,
   donor_name varchar(500) NOT NULL DEFAULT '',
   purchase_date date,
-  denomination integer,
+  denomination bigint,
   p_date varchar(50),
   year smallint,
   month smallint,
